Add tabs to agent settings
