+++ pages/agent/index.js
@@ -1,12 +1,6 @@
 import { useRouter } from "next/router";
 import { useMemo } from "react";
-import useSWR from "swr";
 import AgentPanel from "../../components/systems/agent/AgentPanel";
-import ContentSWR from "../../components/data/ContentSWR";
-import { sdk } from "../../lib/apiClient";
-import ReactMarkdown from "react-markdown";
-import { Container } from "@mui/material";
-import axios from "axios";
 
 export default function Agents({
   chains,
@@ -25,16 +19,6 @@ export default function Agents({
   setUseSelectedAgent,
   theme,
 }) {
-  const docs = useSWR(
-    "docs/agent",
-    async () =>
-      (
-        await axios.get(
-          "https://raw.githubusercontent.com/Josh-XT/AGiXT/main/docs/2-Concepts/9-Agent%20Interactions.md"
-        )
-      ).data
-  );
-
   const router = useRouter();
   const agentName = useMemo(() => router.query.agent, [router.query.agent]);
   if (agentName) {
@@ -58,13 +42,6 @@ export default function Agents({
       />
     );
   } else {
-    return (
-      <Container>
-        <ContentSWR
-          swr={docs}
-          content={({ data }) => <ReactMarkdown>{data}</ReactMarkdown>}
-        />
-      </Container>
-    );
+    router.push("/");
   }
 }
